\name{Chem197}
\alias{Chem197}
\docType{data}
\title{
Chemical Concentrations
}
\description{
Readings of a chemical process every two hours
}
\usage{data("Chem197")}
\format{
  A data frame with 197 observations on the following 2 variables.
  \describe{
    \item{\code{t}}{Time index (1 to 197)}
    \item{\code{Chem}}{Chemical concentration}
  }
}
\details{
Series A from the Box/Jenkins Time Series book.  Chemical concentration readings taken every two hours to get a time series of 197 values.
}
\source{
page 525 - Box/Jenkins "Time Series: Forecasting and Control"
}

\keyword{datasets}
