\name{CPI}
\alias{CPI}
\docType{data}
\title{
Comsumer Price Index
}
\description{
Meidcal and Recreation components of the CPI
}
\usage{data("CPI")}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (1991 to 2020)}
    \item{\code{Medical}}{Medical component of the CPI}
    \item{\code{Recreation}}{Recreation component of the CPI}
  }
}
\details{
Yearly medical and recreation components of the US Consumer Priced Index from 1991 to 2020.
}
\source{
Downloaded from the Bureau of Labor statistics at https://www.bls.gov/CPI/#data
}

\keyword{datasets}
