\name{CPIChange}
\alias{CPIChange}
\docType{data}
\title{
Yearly Change in CPI (1914-2018)
}
\description{
Consumer Price Index (CPI) change each year from 1914 to 2018
}
\usage{data("CPIChange")}
\format{
  A data frame with 105 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (1914 to 2018)}
    \item{\code{Change}}{Change in the Consumer Price Index}
    \item{\code{t}}{Time index (1 to 105)}
  }
}
\details{
Annual changes in the U.S. Consumer Price Index for the years 1914 through 2018.
}
\source{
http://www.bls.gov/cpi/data.htm
}

\keyword{datasets}
