\name{Crude24}
\alias{Crude24}
\docType{data}
\title{
Monthly Crude Oil Prices (2019-2020)
}
\description{
Average monthly price for crude oil (2019-2020)
}
\usage{data("Crude24")}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{Month}}{Month-Year}
    \item{\code{Price}}{Crude oil price (in $ per barrel)}
  }
}
\details{
Average monthly price for crude oil in dollars per barrel for 2019 and 2020.
}
\source{
U.S. Energy Information Administration at https://www.eia.gov/petroleum/data.php#prices
}

\keyword{datasets}
