\name{FrazerMonthly}
\alias{FrazerMonthly}
\docType{data}
\title{
Frazer Computing Sales (Monthly)
}
\description{
Monthly sales at Frazer Computing (1993-2020)
}
\usage{data("FrazerMonthly")}
\format{
  A data frame with 336 observations on the following 3 variables.
  \describe{
    \item{\code{Month}}{Month}
    \item{\code{Year}}{Year}
    \item{\code{Sales}}{Software units sold}
  }
}
\details{
Frazer Computing is a software company that specializes in programs to support car dealers.  The dataset is a monthly time series on units sold since 1993.
}
\source{
Thanks to Michael Frazer, founder of Frazer Computing, for supplying the sales data.
}

\keyword{datasets}
