\name{FrazerYearly}
\alias{FrazerYearly}
\docType{data}
\title{
Frazer Computing Sales (Yearly)
}
\description{
Yearly sales at Frazer Computing (1993-2020)
}
\usage{data("FrazerYearly")}
\format{
  A data frame with 26 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (1993 to 2020)}
    \item{\code{TotalSales}}{Software unit sold}
    \item{\code{t}}{Time index (1 to 28)}
  }
}
\details{
Frazer Computing is a software company that specializes in programs to support car dealers.  The dataset is a time series on units sold each year since 1993
}
\source{
hanks to Michael Frazer, founder of Frazer Computing, for supplying the sales data.
}

\keyword{datasets}
