% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdelta.R
\name{getdelta}
\alias{getdelta}
\title{Compute the estimate for the constant term (delta) when fitting  ARIMA(p,0,q)}
\usage{
getdelta(armamodel)
}
\arguments{
\item{armamodel}{The result of an ARIMA(p,0,q) fit with constant term included}
}
\value{
The estimated value for delta, he constnat term in the ARIMA(p,0,q) model
}
\description{
Compute the estimate for the constant term (delta) when fitting  ARIMA(p,0,q)
}
