\name{GoldenGate}
\alias{GoldenGate}
\docType{data}
\title{
Golden Gate Bridge Traffic (1976-1980)
}
\description{
Monthly traffic (in thousands) over the Golden Gate bridge
}
\usage{data("GoldenGate")}
\format{
  A data frame with 60 observations on the following 4 variables.
  \describe{
    \item{\code{t}}{Time index (1 to 60)}
    \item{\code{traffic}}{Number of vehicles (in thousands)}
    \item{\code{Month}}{Numeric code for month (1 to 12)}
    \item{\code{Monthyr}}{Month and year coded like \code{Aug76}}
  }
}
\details{
Monthly traffic (in thousands) crossing the Golden Gate bridge near San Francisco from 1976 to 1980.
}
\source{
??
}

\keyword{datasets}
