\name{GoldenGate14}
\alias{GoldenGate14}
\docType{data}
\title{
Golden Gate Bridge Traffic (1967-1980)
}
\description{
Monthly traffic (in thousands) over the Golden Gate bridge over 14 years
}
\usage{data("GoldenGate14")}
\format{
  A data frame with 168 observations on the following 5 variables.
  \describe{
    \item{\code{traffic}}{Number of vehicles (in thousands)}
    \item{\code{date}}{Month and year coded like \code{Aug-76}}
    \item{\code{t}}{Time index (1 to 168)}
    \item{\code{month}}{Numeric code for month (1 to 12)}
    \item{\code{year}}{Year (1967 to 1980)}
  }
}
\details{
Monthly traffic (in thousands) crossing the Golden Gate bridge near San Francisco from 1967 to 1980.
}
\source{
???
}

\keyword{datasets}
