\name{HouseStarts1979}
\alias{HouseStarts1979}
\docType{data}
\title{
Monthly U.S. Housing Starts (1977-1979)
}
\description{
Monthly U.S. housing starts  from 1977 though 1979.
}
\usage{data("HouseStarts1979")}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (1977-1979)}
    \item{\code{Qtr}}{Code for quarter (1 to 4)}
    \item{\code{Houses}}{Housing starts each quarter (in thousands)}
  }
}
\details{
Data give the number of new house starts in the U.S. for each quarter for three years (1977 to 1979).
}
\source{
Statistics for Business and Economics by McClave & Benson   (pg. 615)
}

\keyword{datasets}
