\name{HouseStarts2005}
\alias{HouseStarts2005}
\docType{data}
\title{
Monthly U.S. Housing Starts (2000-2005)
}
\description{
Monthly U.S. housing starts  from 2000 though 2005.
}
\usage{data("HouseStarts2005")}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (2000 to 2005)}
    \item{\code{Qtr}}{Code for quarter (1 to 4)}
    \item{\code{Starts}}{Housing starts each quarter (in thousands)}
  }
}
\details{
Data give the number of new house starts in the U.S. for each quarter for six years (2000 to 2005).
}
\source{
https://www.census.gov/construction/nrc/index.html
}

\keyword{datasets}
