\name{HouseStarts2018}
\alias{HouseStarts2018}
\docType{data}
\title{
Monthly U.S. Housing Starts (2000-2018)
}
\description{
Monthly U.S. housing starts  from 2000 though 2018.
}
\usage{data("HouseStarts2018")}
\format{
  A data frame with 76 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (2000 to 2018)}
    \item{\code{Qtr}}{Code for quarter (1 to 4)}
    \item{\code{Starts}}{Housing starts each quarter (in thousands)}
  }
}
\details{
Data give the number of new house starts in the U.S. for each quarter for 14 years (2000 to 2018).
}
\source{
https://www.census.gov/construction/nrc/index.html
}

\keyword{datasets}
