\name{HouseStarts2020}
\alias{HouseStarts2020}
\docType{data}
\title{
Monthly U.S. Housing Starts (2000-2020)
}
\description{
Monthly U.S. housing starts  from 2000 though 2020.
}
\usage{data("HouseStarts2020")}
\format{
  A data frame with 83 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (2000 to 2020)}
    \item{\code{Qtr}}{Code for quarter (1 to 4)}
    \item{\code{Starts}}{Housing starts each quarter (in thousands)}
  }
}
\details{
Data give the number of new house starts in the U.S. for each quarter for 21 years (2000 to 2020). Missing Q4 of 2020.
}
\source{
https://www.census.gov/construction/nrc/index.html
}

\keyword{datasets}
