\name{HouseStarts22}
\alias{HouseStarts22}
\docType{data}
\title{
Monthly U.S. Housing Starts (2000-2022)
}
\description{
Monthly U.S. housing starts  from 2000 though 2022.
}
\usage{data("HouseStarts22")}
\format{
  A data frame with 91 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (2000 to 2022)}
    \item{\code{Qtr}}{Code for quarter (1 to 4)}
    \item{\code{Starts}}{Housing starts each quarter (in thousands)}
  }
}
\details{
Data give the number of new house starts in the U.S. for each quarter for 23 years (2000 to 2022). Missing Q4 of 2022.
}
\source{
https://www.census.gov/construction/nrc/index.html (New Residential Construction > Housing Units Started > Single-family Units > Not Seasonally Adjusted)
}

\keyword{datasets}
