\name{IBM12}
\alias{IBM12}
\docType{data}
\title{
IBM Stock Prices (12 days)
}
\description{
Closing price of IBM stock for 12 days
}
\usage{data("IBM12")}
\format{
  A data frame with 12 observations on the following variable.
  \describe{
    \item{\code{Price}}{Closing IBM stock price}
  }
}
\details{
This is a subset of IBM369 with just the first 12 days of daily IBM stock prices starting on May 17, 1961.
}
\source{
page 526 (first 12 values) from Box/Jenkins "Time Series: Forecasting and Control"
}

\keyword{datasets}
