\name{IBM369}
\alias{IBM369}
\docType{data}
\title{
IBM Stock Prices (369 days)
}
\description{
Closing price of IBM stock for 369 days
}
\usage{data("IBM369")}
\format{
  A data frame with 369 observations on the following variable.
  \describe{
    \item{\code{IBM}}{Closing IBM Stock priced}
  }
}
\details{
Daily price of IBM stock starting in May 5, 1961,  This is Series B in the Box/Jenkins Time Series book.
}
\source{
page 526 - Box/Jenkins "Time Series: Forecasting and Control"
}

\keyword{datasets}
