\name{MilkProd}
\alias{MilkProd}
\docType{data}
\title{
US Milk Production
}
\description{
U.S> Milk Productio (monthly 2020-2022)
}
\usage{data("MilkProd")}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year (2020-2022)}
    \item{\code{Period}}{Month (JAN, FEB, ...)}
    \item{\code{Month}}{Month (1, 2, ...)}
    \item{\code{Milk}}{Milk production ( billions of gallons)}
  }
}
\details{
MOnthly milk production in the U.S. over three years (2020 to 2022)
}
\source{
Downloaded from the USDA  https://quickstats.nass.usda.gov/
}

\keyword{datasets}
