\name{PeaceBridge2003}
\alias{PeaceBridge2003}
\docType{data}
\title{
Monthly Traffic over the Peace Bridge (2003-2018)
}
\description{
Monthly traffic (in thousands) over the Peace Bridge (2003-2018)
}
\usage{data("PeaceBridge2003")}
\format{
  A data frame with 192 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year (2003 to 2018)}
    \item{\code{Month}}{Numeric code for the month (1 to 12)}
    \item{\code{Traffic}}{Number of vehicles (in thousands)}
    \item{\code{t}}{Time index (1 to 192)}
  }
}
\details{
The data record the number of vehicles crossing the Peace Bridge (between the US and Canada near Niagara Falls) each month for 16 years (2003 to 2018)
}
\source{
https://www.peacebridge.com/index.php/historical-traffic-statistics/yearly-volumes
}

\keyword{datasets}
