\name{PeaceBridge2012}
\alias{PeaceBridge2012}
\docType{data}
\title{
Monthly Traffic over the Peace Bridge (2012-2015)
}
\description{
Monthly traffic (in thousands) over the Peace Bridge (2012-2015)
}
\usage{data("PeaceBridge2012")}
\format{
  A data frame with 48 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year (2012 to 2015)}
    \item{\code{Month}}{Numeric code for the month (1 to 12)}
    \item{\code{Traffic}}{Number of vehicles (in thousands)}
    \item{\code{t}}{Time index (1 to 48)}
  }
}
\details{
The data record the number of vehicles crossing the Peace Bridge (between the US and Canada near Niagara Falls) each month for 4 years (2012 to 2015)
}
\source{
https://www.peacebridge.com/index.php/historical-traffic-statistics/yearly-volumes
}

\keyword{datasets}
