\name{RetailGas}
\alias{RetailGas}
\docType{data}
\title{
Monthly Retail Gas Prices
}
\description{
Average retail gas prices in the US (Sept2022-Aug 2024)
}
\usage{data("RetailGas")}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{Month}}{Month-Year}
    \item{\code{Price}}{Retai gas price in dollars}
  }
}
\details{
Mean monthly price for all grades of gasoline in the United States from Sepember 2022 to August 2024.
}
\source{
U.S. Energy Information Administration at https://www.eia.gov/petroleum/gasdiesel/
}

\keyword{datasets}
