\name{RetailSales18}
\alias{RetailSales18}
\docType{data}
\title{
Monthly Retail Sales (2000-2018)
}
\description{
U.S. retail sales (in billions of dollars) from Jan-00 to Nov-18.
}
\usage{data("RetailSales18")}
\format{
  A data frame with 228 observations on the following 4 variables.
  \describe{
    \item{\code{Time}}{Month-Year (e.g. Jan-03)}
    \item{\code{Sales}}{U.S. Retail sales (in billions)}
    \item{\code{Month}}{Coded month (Jan=1 to Dec=12)}
    \item{\code{t}}{Time index (1 to 228)}
  }
}
\details{
Monthly U.S. retail sales (in billions) from 2000-2018.  This differs from RetailSales in that only December 2018 is missing.
}
\source{
Data from the U.S. Census website at https://www.census.gov/retail/index.html#mrts.
}

\keyword{datasets}
