\name{RetailSales20}
\alias{RetailSales20}
\docType{data}
\title{
Monthly Retail Sales (2009-2020)
}
\description{
U.S. retail sales (in billions of dollars) from Jan-2009 to Sept-2020.
}
\usage{data("RetailSales20")}
\format{
  A data frame with 141 observations on the following 4 variables.
  \describe{
    \item{\code{Time}}{Month-Year (e.g. Jan-2009)}
    \item{\code{Sales}}{U.S. Retail sales (in billions)}
    \item{\code{Month}}{Coded month (Jan=1 to Dec=12)}
    \item{\code{t}}{Time index (1 to 119)}
  }
}
\details{
Monthly U.S. retail sales (in billions) from 2009-2020.  This differs from RetailSales in that it starts in 2009 an goes through September 2020.
}
\source{
Data from the U.S. Census website at https://www.census.gov/retail/index.html#mrts.
}

\keyword{datasets}
