\name{RetailSales22}
\alias{RetailSales22}
\docType{data}
\title{
Monthly Retail Sales (2009-2022)
}
\description{
U.S. retail sales (in billions of dollars) from Jan-2009 to Sept-2022.
}
\usage{data("RetailSales22")}
\format{
  A data frame with 165 observations on the following 4 variables.
  \describe{
    \item{\code{Time}}{Month-Year (e.g. Jan-2009)}
    \item{\code{Sales}}{U.S. Retail sales (in billions)}
    \item{\code{Month}}{Coded month (Jan=1 to Dec=12)}
    \item{\code{t}}{Time index (1 to 165)}
  }
}
\details{
Monthly U.S. retail sales (in billions) from 2009-2022.  This differs from RetailSales in that it starts in 2009 an goes through September 2022.
}
\source{
Data from the U.S. Census website at https://www.census.gov/retail/index.html#mrts.
}

\keyword{datasets}
