% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runavg.R
\name{runavg}
\alias{runavg}
\title{Create a running average of past values for a historical time series.}
\usage{
runavg(z, k, na.rm = TRUE)
}
\arguments{
\item{z}{A time series or vector of values}

\item{k}{Number of previous terms to average}

\item{na.rm}{Remove the inital NA values (default is TRUE)}
}
\value{
A time series (or vector) of the averages taken k at a time with the first value at time k+1
}
\description{
Create a running average of past values for a historical time series.
}
