% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesplot.R
\name{sesplot}
\alias{sesplot}
\title{Plot forecasts (historical and future) for exponential smoothing models}
\usage{
sesplot(mod, include = NULL, attach = "fits", main = "", lwd = 2)
}
\arguments{
\item{mod}{The result of ftting an exponential smoothing moel with ses( ), holt( ) or hw( )}

\item{include}{A number of recnt past values to show (default is NULL for all)}

\item{attach}{Attach the future forecasts (and bounds) the the historical "fits" (default), or historical "series", or "no" attaching}

\item{main}{Usual paramter to enter a title for the graph}

\item{lwd}{Control line width (default is lwd=2)}
}
\value{
A plot with the series, historical (smoothed) values, forecast (and bounds) for future values.
}
\description{
Plot forecasts (historical and future) for exponential smoothing models
}
