% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAR1.R
\name{simAR1}
\alias{simAR1}
\title{Simulate a First order Autoregressive Model AR(1)}
\usage{
simAR1(n = 100, delta = 0, phi = 0.5, sderror = 1)
}
\arguments{
\item{n}{Length of the series to simulate}

\item{delta}{Constant term (delta)}

\item{phi}{Coefficient of Z(t-1)}

\item{sderror}{Standard deviation of the error term}
}
\value{
A time series simulated from the model
}
\description{
Simulate a First order Autoregressive Model AR(1)
}
