% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simARIMA.R
\name{simARIMA}
\alias{simARIMA}
\title{Simulate data from an ARIMA model}
\usage{
simARIMA(n = 100, delta = 0, phi = NULL, d = 0, theta = NULL,
  sderror = 1)
}
\arguments{
\item{n}{Length of the series to simulate}

\item{delta}{Constant term (delta)}

\item{phi}{AR coefficients (use c(  ) to enter more than one)}

\item{d}{Number of differences (default is d=0)}

\item{theta}{Moving average coefficients (use c( ) to enter more than one)}

\item{sderror}{Standard deviation of the error term}
}
\value{
A time series simulated from the ARIMA(p,d,q) model
}
\description{
Simulate data from an ARIMA model
}
