% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simModel3.R
\name{simModel3}
\alias{simModel3}
\title{Simulate an MA(1) model (Model #3)}
\usage{
simModel3(n = 100, mean = 0, theta = 0.5, sderror = 1)
}
\arguments{
\item{n}{Length of the series to simulate}

\item{mean}{Mean (theoretical) of the series}

\item{theta}{Coeffiecient of error a(t-1)}

\item{sderror}{Standard deviation of the error term}
}
\value{
A time series simulated from the model
}
\description{
Simulate an MA(1) model (Model #3)
}
