% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simWalk.R
\name{simWalk}
\alias{simWalk}
\title{Simulate a Random Walk}
\usage{
simWalk(n = 100, z0 = 0, drift = 0, sderror = 1)
}
\arguments{
\item{n}{Length of the series to simulate}

\item{z0}{Starting point for the series}

\item{drift}{Value for a drift parameter (default is zero)}

\item{sderror}{Standard deviation of the error term}
}
\value{
A time series simulated from the model
}
\description{
Simulate a Random Walk
}
