% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sluacf.R
\name{sluacf}
\alias{sluacf}
\title{Compute the autocorrelation of a time series}
\usage{
sluacf(
  series,
  maxlag = NULL,
  plot = TRUE,
  ndiff = 0,
  sdiff = 0,
  out = TRUE,
  ci.type = "ma"
)
}
\arguments{
\item{series}{A time series}

\item{maxlag}{Number of autocorrelations to return (Null default picks a reasoanble number)}

\item{plot}{Plot of the ACF with bounds? (default is TRUE)}

\item{ndiff}{Number of regualr differences (default is 0)}

\item{sdiff}{Number of seasonal differences (default is zero)}

\item{out}{Output the numerical ACF? (default is TRUE)}

\item{ci.type}{Type of confidence bounds ("ma" defualt or "white" for white noise)}
}
\value{
An autocorrelation object with the first component(lag 0) set to NA
}
\description{
Compute the autocorrelation of a time series
}
