% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slunova.R
\name{slunova}
\alias{slunova}
\title{Print a more standard ANOVA table for a regression model}
\usage{
slunova(model)
}
\arguments{
\item{model}{Result from an lm( ) regression mdoel}
}
\description{
Print a more standard ANOVA table for a regression model
}
