% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slupacf.R
\name{slupacf}
\alias{slupacf}
\title{Compute the autocorrelation of a time series}
\usage{
slupacf(series, maxlag = NULL, plot = FALSE, ndiff = 0, sdiff = 0,
  out = TRUE, ci.type = "ma")
}
\arguments{
\item{series}{A time series}

\item{maxlag}{Number of partialautocorrelations to return (Null default picks a reasoanble number)}

\item{plot}{Plot of the PACF with bounds? (default is FALSE)}

\item{ndiff}{Number of regular differences (default is 0)}

\item{sdiff}{Number of seasonal differences (default is zero)}

\item{out}{Output the numerical ACF? (default is TRUE)}

\item{ci.type}{Type of confidence bounds ("ma" defualt or "white" for white noise)}
}
\value{
An partial autocorrelation object
}
\description{
Compute the autocorrelation of a time series
}
