\name{StatKeyF2024}
\alias{StatKeyF2024}
\docType{data}
\title{
StatKey Sessions
}
\description{
Daily StatKey sessions (9/1/2024 to 9/7/2024)
}
\usage{data("StatKeyS2024")}
\format{
  A data frame with 98 observations on the following 2 variables.
  \describe{
    \item{\code{Date}}{Day with format mm/dd/yyy}
    \item{\code{Users}}{Number of StatKey users}
    \item{\code{t}}{Time index 1 to 98}
    \item{\code{Day}}{Day coded as Sun, Mon, Tue, Wed, Thu, Fri, or Sat}

  }
}
\details{
Daily sessions of the online StatKey software (lock5stat.com/StatKey) for 14 weeks () during the fall semester of 2024.
}
\source{
Data obtained via Google Analytics by the StatKey developers.
}

\keyword{datasets}
