\name{StatKeyS2019}
\alias{StatKeyS2019}
\docType{data}
\title{
StatKey Users
}
\description{
Daily StatKey users (1/13/2019 to 4/27/2019)
}
\usage{data("StatKeyS2019")}
\format{
  A data frame with 105 observations on the following 2 variables.
  \describe{
    \item{\code{Date}}{Day with format mm/dd/yyy}
    \item{\code{Users}}{Number of StatKey users}
  }
}
\details{
Daily users of the online StatKey software (lock5stat.com/StatKey) for 15 weeks (1/13/2019 to 4/27/2019) during the spring semester of 2019.
}
\source{
Data obtained via Google Analytics by the StatKey developers.
}

\keyword{datasets}
