\name{StatKeyS2021}
\alias{StatKeyS2021}
\docType{data}
\title{
StatKey Users
}
\description{
Daily StatKey users (1/4/2021 to 4/3/2021)
}
\usage{data("StatKeyS2021")}
\format{
  A data frame with 105 observations on the following 2 variables.
  \describe{
    \item{\code{Date}}{Day with format mm/dd/yyy}
    \item{\code{Users}}{Number of StatKey users}
    \item{\code{Day}}{Day coded as Su, M, Tu, W, Th, F, or Sa}
    \item{\code{t}}{Time index 1 to 91}
  }
}
\details{
Daily users of the online StatKey software (lock5stat.com/StatKey) for 13 weeks (1/4/2021 to 4/3/2021) during the spring semester of 2021.
}
\source{
Data obtained via Google Analytics by the StatKey developers.
}

\keyword{datasets}
