\name{StatKeyS2023}
\alias{StatKeyS2023}
\docType{data}
\title{
StatKey Sessions
}
\description{
Daily StatKey sessions (1/22/2023 to 4/29/2022)
}
\usage{data("StatKeyS2023")}
\format{
  A data frame with 98 observations on the following 2 variables.
  \describe{
    \item{\code{Date}}{Day with format mm/dd/yyy}
    \item{\code{Users}}{Number of StatKey sessions}
    \item{\code{t}}{Time index 1 to 98}
    \item{\code{Day}}{Day coded as Sun, Mon, Tue, Wed, Thu, Fri, or Sat}

  }
}
\details{
Daily sessions of the online StatKey software (lock5stat.com/StatKey) for 14 weeks () during the spring semester of 2023.
}
\source{
Data obtained via Google Analytics by the StatKey developers.
}

\keyword{datasets}
