\name{SuperBowl}
\alias{SuperBowl}
\docType{data}
\title{
Super Bowl Scores
}
\description{
Winner's and loser's score and margin of victory for Super Bowls from 1967 to 2020
}
\usage{data("SuperBowl")}
\format{
  A data frame with 54 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year (1967 to 2020)}
    \item{\code{Winner}}{Points scored by the winning team}
    \item{\code{Loser}}{Pooints scored by the losing team}
    \item{\code{Diff}}{Margin if victory (winner - loser)}
  }
}
\details{
Scores from past Super Bowl (American football) games.  For the years 1967 to 2020, the dataset has the winning and losing scores as well as the margin of victory.
}
\source{
Scores are available at lots of online sources, such as https://www.pro-football-reference.com/super-bowl/
}

\keyword{datasets}
