% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testacf.R
\name{testacf}
\alias{testacf}
\title{Provide test statistics for autocorrelations (individual and LBQ)}
\usage{
testacf(seriesacf)
}
\arguments{
\item{seriesacf}{The result of an autocorreltion calculation, e.g. from acf( ) or sluacf( )}
}
\value{
A dataframe with lag, autocorrelation (rk), std. error of rk (Srk), test statistic (tsrk) and Ljung-Box statistic (LBQ)
}
\description{
Provide test statistics for autocorrelations (individual and LBQ)
}
