% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpacf.R
\name{testpacf}
\alias{testpacf}
\title{Provide test statistics for partial autocorrelations}
\usage{
testpacf(seriespacf)
}
\arguments{
\item{seriesacf}{The result of a partial autocorreltion calculation, e.g. from pacf( ) or slupacf( )}
}
\value{
A dataframe with lag, partial autocorrelation (phikk), std. error of phikk (Sphikk), test statistic (tsphikk) and Ljung-Box statistic (LBQ)
}
\description{
Provide test statistics for partial autocorrelations
}
