\name{USBeer}
\alias{USBeer}
\docType{data}
\title{
U.S. Beer Production (Quarterly)
}
\description{
Quarterly U.S> Beer production from 1990 to 2022
}
\usage{data("USBeer")}
\format{
  A data frame with 130 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (1990 to 2022)}
    \item{\code{Qtr}}{Quarer (coded as 1, 2, 3, or 4)}
    \item{\code{Beer}}{U.S. beer production (in millions of barrels)}
  }
}
\details{
Quarterly procduction of beer in the United States from 1990 to 2022.  The last two quarters of 2022 are missing.
}
\source{
Data obtained from the U.S. Alochol and Tobacco Tax and Trade Bureau https://www.ttb.gov/beer/beer-stats.shtml
}

\keyword{datasets}
