\name{Wages}
\alias{Wages}
\docType{data}
\title{
Average Hourly Wage
}
\description{
Yealy data (1972-2022) on the average hourly wage for workers in three employment sectors: Durable goods, Wholesale sales, and retail sales
}
\usage{data("Wages")}
\format{
  A data frame with 51 observations on the following 4 variables.
  \describe{
    \item{\code{Year}}{Year (1972 to 2022)}
    \item{\code{Durable}}{Average hourly wage (in dollars) for workers in the durable goods sector}
    \item{\code{Wholesale}}{Average hourly wage (in dollars) for workers in the wholesale sales sector}
    \item{\code{Retail}}{Average hourly wage (in dollars) for workers in the retail sales sector}
  }
}
\details{
The dataset gives the average hourly wages each year for three sectors of employment: durable goods, wholesale sales, and retail sales.
}
\source{
Data obtained from the Bureau of Labor Statistics at https://data.bls.gov/PDQWeb/ce
}

\keyword{datasets}
