% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyChar.R
\name{polyChar}
\alias{polyChar}
\title{Show characteristic polynomials from an Arima model}
\usage{
polyChar(model)
}
\arguments{
\item{model}{an object of class “arima”. For example, model produced by the Arima() function}
}
\value{
A list of characteristic polynomials for a given fitted ARIMA model
}
\description{
Show characteristic polynomials from an Arima model
}
