\name{SLUGifts}
\alias{SLUGifts}
\docType{data}
\title{
Annual Giving at St. Lawrence University (1996-2023)
}
\description{
Yearly value of gifts to St. Lawrence Univeristy as recorded by the Development Office.
}
\usage{data("SLUGifts")}
\format{
  A data frame with 28 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{Year (1996-2023)}
    \item{\code{t}}{Time index (1 to 28)}
    \item{\code{Gifts}}{Total value of annual gifts (in millions)}
  }
}
\details{
The Development Office at St. Lawrence University reports annual giving to the university.  Data were recorded from 1996 through 2023.
}
\source{
Data obtained from St. Lawrence University, http://www.stlawu.edu
}

\keyword{datasets}
