% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpacf.R
\name{testpacf}
\alias{testpacf}
\title{Provide test statistics for partial autocorrelations}
\usage{
testpacf(seriespacf)
}
\arguments{
\item{seriespacf}{The result of a partial autocorreltion calculation, e.g. from pacf( ) or slupacf( )}
}
\value{
A dataframe with lag, partial autocorrelation (rkk), std. error of rkk (Srkk), test statistic (tsrkk)
}
\description{
Provide test statistics for partial autocorrelations
}
