\name{USVehicles}
\alias{USVehicles}
\docType{data}
\title{
U.S. Vehicle Registrations (Yearly)
}
\description{
Yearly count of number of motor vehicles registered in the U.S. (1948-2021)
}
\usage{data("USVehicles")}
\format{
  A data frame with 74 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{Year (1948-2021)}
    \item{\code{Vehicles}}{Number of registered motor vehicles (in millions)}
  }
}
\details{
The dataset has the number of motor vehicles (in millions) registered in the Uniged States for each year from 1948 to 2021.
}
\source{
For recent years used Google public data (https://www.google.com/publicdata/explore?ds=gb66jodhlsaab_) to access data from the Office of Highway Policy Information, https://www.fhwa.dot.gov/policyinformation/index.cfm.  Also from https://www.fhwa.dot.gov/policyinformation/statistics/2019/mv1.cfm
}

\keyword{datasets}
